requirements.txt
djangorestframework-simplejwt==4.8.0

djangorestframework==3.13.1

django-rest-authtoken==2.1.4

Django==4.0.4

drf-yasg==1.21.3